LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/reporting-observer-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: ReportingObserver

Harness Error (TIMEOUT), message = null

TIMEOUT Opener COOP Test timed out
NOTRUN Openee COOP
NOTRUN Access from same-origin iframe
NOTRUN Access from cross-site iframe
